move slider to atoms
