+++ app/src/atoms/Slider/index.tsx
