+++ app/src/atoms/Slider/slider.module.css
